functions/reset_warning/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.5.3
